Removed promises.ts
