+++ src/main/internal/index.ts
@@ -2,6 +2,5 @@ export * from './arrays';
 export * from './bitmasks';
 export * from './lang';
 export * from './objects';
-export * from './promises';
 export * from './shapes';
 export * from './types';
